backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
newspaper3k
lxml[html_clean]
